split out rtc device
